requirement.txt
glcontext==3.0.0
moderngl==5.12.0
moderngl-window==3.1.1
numpy==2.3.2
pillow==11.3.0
psutil==7.0.0
pygame==2.6.1
pyglet==2.1.8
pyglm==2.8.2
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
